Remove ManagerOptions test and add ListenerOptions test
